Fixes gaps in shadows and shadow update jitters.
Gaps in shadows would appear when parts of objects behind the camera were missing from the shadow map. Also, shadow edges would jitter when rotating the camera. The solution for both of these issues was to use a uniform box, centered at the camera's position, instead of tightly fitting a non-uniform box to the camera's frustrum.